fastapi/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
anthropic
pymongo>=4.6
